pyproject.toml
[project]
name = "cv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prek>=0.3.8",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "typer>=0.24.1",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]

[project.scripts]
cv-build = "cv.run:app"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["tests*"]
